+++ QuickTableViewControllerTests/SwitchRowSpec.swift
@@ -28,7 +28,7 @@ import Nimble
 import Quick
 import QuickTableViewController
 
-class SwitchRowSpec: QuickSpec {
+internal final class SwitchRowSpec: QuickSpec {
 
   override func spec() {
     describe("initialization") {
@@ -38,7 +38,7 @@ class SwitchRowSpec: QuickSpec {
       it("should initialize with given parameters") {
         expect(row.title) == "title"
         expect(row.switchValue) == true
-        expect(row.cellReuseIdentifier) == "QuickTableViewController.SwitchCell"
+        expect(row.cellReuseIdentifier) == "SwitchCell"
         expect(row.action).notTo(beNil())
 
         row.action?(row)
@@ -82,9 +82,10 @@ class SwitchRowSpec: QuickSpec {
       var toggled = false
       var row = SwitchRow(title: "", switchValue: false) { _ in toggled = !toggled }
 
-      it("should invoke the action closure when the value changes") {
+      // The action invocation has moved to the view controller.
+      it("should not invoke the action closure when the value changes") {
         row.switchValue = true
-        expect(toggled) == true
+        expect(toggled) == false
       }
     }
   }